Make EdgeHTML open source
As another big step I would like to see EdgeHTML open sourced just as MS did with the ChakraCore. ChakraCore is a rip-off from the IE just as EdgeHTML is a refreshed version of Trident from IE. So why not? Don't see this as real pull request, more like a christmas wish :smile: